requirements.txt
pre-commit
yapf
pylint

torch==1.10.0
torchvision==0.11.1
numpy
Pillow
tqdm
matplotlib
pyyaml
wandb
opencv-python
colour_demosaicing
scipy
sklearn
scikit-image
